+++ braintree/client_token.py
@@ -20,8 +20,6 @@ class ClientToken(object):
         return [
             "customer_id",
             "proxy_merchant_id",
-            "sepa_mandate_type",
-            "sepa_mandate_acceptance_location",
             "version",
             "merchant_account_id",
             {"options": ["make_default", "verify_card", "fail_on_duplicate_payment_method"]}