+++ tests/integration/test_us_bank_account.py
@@ -5,7 +5,10 @@ class TestUsBankAccount(unittest.TestCase):
         customer_id = Customer.create().customer.id
         result = PaymentMethod.create({
             "customer_id": customer_id,
-            "payment_method_nonce": TestHelper.generate_valid_us_bank_account_nonce()
+            "payment_method_nonce": TestHelper.generate_valid_us_bank_account_nonce(),
+            "options": {
+                "verification_merchant_account_id": "us_bank_merchant_account"
+            }
         })
         self.assertTrue(result.is_success)
 
@@ -27,7 +30,10 @@ class TestUsBankAccount(unittest.TestCase):
         customer_id = Customer.create().customer.id
         result = PaymentMethod.create({
             "customer_id": customer_id,
-            "payment_method_nonce": TestHelper.generate_valid_us_bank_account_nonce()
+            "payment_method_nonce": TestHelper.generate_valid_us_bank_account_nonce(),
+            "options": {
+                "verification_merchant_account_id": "us_bank_merchant_account"
+            }
         })
         self.assertTrue(result.is_success)